use yan dup_osc_idx_iter
